+++ react/javascript/test/components/customise/CustomRenderingTest.tsx
@@ -1,19 +1,12 @@
 import assert from 'assert'
-import ReactDOM from 'react-dom'
 import React from 'react'
-import { JSDOM } from 'jsdom'
 import Tags from '../../../src/components/gherkin/Tags'
 import CustomRendering from '../../../src/components/customise/CustomRendering'
 import * as messages from '@cucumber/messages'
+import { render } from '../utils'
 
 describe('custom rendering and theming', () => {
   it('uses the generated class names from built-in styles by default', () => {
-    const dom = new JSDOM('<html lang="en"><body><div id="content"></div></body></html>')
-    // @ts-ignore
-    global.window = dom.window
-    // global.navigator = dom.window.navigator
-    const document = dom.window.document
-
     const tags: messages.Tag[] = [
       {
         id: '123',
@@ -25,20 +18,13 @@ describe('custom rendering and theming', () => {
       },
     ]
 
-    const app = <Tags tags={tags} />
-    ReactDOM.render(app, document.getElementById('content'))
+    const { container } = render(<Tags tags={tags} />)
 
-    assert.equal(document.querySelector('ul').className, 'tags__generated')
-    assert.equal(document.querySelector('ul > li').className, 'tag__generated')
+    assert.strictEqual(container.querySelector('ul').className, 'tags__generated')
+    assert.strictEqual(container.querySelector('ul > li').className, 'tag__generated')
   })
 
   it('uses the custom classnames provided via custom rendering', () => {
-    const dom = new JSDOM('<html lang="en"><body><div id="content"></div></body></html>')
-    // @ts-ignore
-    global.window = dom.window
-    // global.navigator = dom.window.navigator
-    const document = dom.window.document
-
     const tags: messages.Tag[] = [
       {
         id: '123',
@@ -50,7 +36,7 @@ describe('custom rendering and theming', () => {
       },
     ]
 
-    const app = (
+    const { container } = render(
       <CustomRendering
         support={{
           Tags: {
@@ -62,19 +48,12 @@ describe('custom rendering and theming', () => {
         <Tags tags={tags} />
       </CustomRendering>
     )
-    ReactDOM.render(app, document.getElementById('content'))
 
-    assert.equal(document.querySelector('ul').className, 'custom-list-class')
-    assert.equal(document.querySelector('ul > li').className, 'custom-item-class')
+    assert.strictEqual(container.querySelector('ul').className, 'custom-list-class')
+    assert.strictEqual(container.querySelector('ul > li').className, 'custom-item-class')
   })
 
   it('uses a partial of custom classes and falls back to built-in styles where omitted', () => {
-    const dom = new JSDOM('<html lang="en"><body><div id="content"></div></body></html>')
-    // @ts-ignore
-    global.window = dom.window
-    // global.navigator = dom.window.navigator
-    const document = dom.window.document
-
     const tags: messages.Tag[] = [
       {
         id: '123',
@@ -86,7 +65,7 @@ describe('custom rendering and theming', () => {
       },
     ]
 
-    const app = (
+    const { container } = render(
       <CustomRendering
         support={{
           Tags: {
@@ -97,19 +76,12 @@ describe('custom rendering and theming', () => {
         <Tags tags={tags} />
       </CustomRendering>
     )
-    ReactDOM.render(app, document.getElementById('content'))
 
-    assert.equal(document.querySelector('ul').className, 'custom-list-class')
-    assert.equal(document.querySelector('ul > li').className, 'tag__generated')
+    assert.strictEqual(container.querySelector('ul').className, 'custom-list-class')
+    assert.strictEqual(container.querySelector('ul > li').className, 'tag__generated')
   })
 
   it('uses a custom component implementation where provided', () => {
-    const dom = new JSDOM('<html lang="en"><body><div id="content"></div></body></html>')
-    // @ts-ignore
-    global.window = dom.window
-    // global.navigator = dom.window.navigator
-    const document = dom.window.document
-
     const tags: messages.Tag[] = [
       {
         id: '123',
@@ -125,7 +97,7 @@ describe('custom rendering and theming', () => {
       return <p>Totally custom!</p>
     }
 
-    const app = (
+    const { container } = render(
       <CustomRendering
         support={{
           Tags: CustomComponent,
@@ -134,18 +106,11 @@ describe('custom rendering and theming', () => {
         <Tags tags={tags} />
       </CustomRendering>
     )
-    ReactDOM.render(app, document.getElementById('content'))
 
-    assert.equal(document.getElementById('content').innerHTML, '<p>Totally custom!</p>')
+    assert.strictEqual(container.innerHTML, '<p>Totally custom!</p>')
   })
 
   it('a custom component can defer to the default renderer if it wants to', () => {
-    const dom = new JSDOM('<html lang="en"><body><div id="content"></div></body></html>')
-    // @ts-ignore
-    global.window = dom.window
-    // global.navigator = dom.window.navigator
-    const document = dom.window.document
-
     const tags: messages.Tag[] = [
       {
         id: '123',
@@ -161,7 +126,7 @@ describe('custom rendering and theming', () => {
       return <props.DefaultRenderer {...props} />
     }
 
-    const app = (
+    const { container } = render(
       <CustomRendering
         support={{
           Tags: CustomComponent,
@@ -170,9 +135,8 @@ describe('custom rendering and theming', () => {
         <Tags tags={tags} />
       </CustomRendering>
     )
-    ReactDOM.render(app, document.getElementById('content'))
 
-    assert.equal(document.querySelector('ul').className, 'tags__generated')
-    assert.equal(document.querySelector('ul > li').className, 'tag__generated')
+    assert.strictEqual(container.querySelector('ul').className, 'tags__generated')
+    assert.strictEqual(container.querySelector('ul > li').className, 'tag__generated')
   })
 })