+++ compatibility-kit/javascript/features/attachments/attachments.feature.ts
@@ -46,5 +46,9 @@ When(
 )
 
 When('a JPEG image is attached', async function () {
-  await this.attach(fs.createReadStream(__dirname + '/cucumber-growing-on-vine.jpg'), 'image/jpg')
+  await this.attach(fs.createReadStream(__dirname + '/cucumber.png'), 'image/png')
+})
+
+When('the {word} png is attached', async function (filename) {
+  await this.attach(fs.createReadStream(__dirname + `/${filename}`), 'image/png')
 })