+++ compatibility-kit/ruby/lib/messages_comparator.rb
@@ -55,7 +55,6 @@ module CCK
     end
 
     def compare_message(found, expected)
-      return unless found.is_a?(Protobuf::Message)
       return if found.is_a?(Cucumber::Messages::GherkinDocument)
       return if found.is_a?(Cucumber::Messages::Pickle)
       return if found.is_a?(Cucumber::Messages::Timestamp) & expected.is_a?(Cucumber::Messages::Timestamp)
@@ -70,10 +69,10 @@ module CCK
       return unless expected.respond_to? :to_hash
       expected.to_hash.keys.each do |key|
         value = expected.send(key)
-        if value.is_a?(Protobuf::Message)
-          compare_message(found.send(key), value)
-        elsif value.is_a?(Array)
+        if value.is_a?(Array)
           compare_list(found.send(key), value)
+        else
+          compare_message(found.send(key), value)
         end
       end
     end