+++ messages/jsonschema/scripts/codegen.rb
@@ -3,6 +3,8 @@ require 'erb'
 require 'set'
 
 class Codegen
+  TEMPLATES_DIRECTORY = "#{File.dirname(__FILE__)}/templates/"
+
   def initialize(paths, template, enum_template, language_type_by_schema_type)
     @paths = paths
     @template = ERB.new(template, nil, '-')
@@ -84,7 +86,7 @@ class Codegen
         if enum
           enum_type_name = "#{parent_type_name}#{capitalize(property_name)}"
           @enums.add({ name: enum_type_name, values: enum })
-          enum_type_name
+          property_type_from_enum(enum_type_name)
         else
           @language_type_by_schema_type[type]
         end
@@ -95,6 +97,14 @@ class Codegen
     end
   end
 
+  def property_type_from_ref(ref)
+    class_name(ref)
+  end
+
+  def property_type_from_enum(enum)
+    enum
+  end
+
   def class_name(ref)
     File.basename(ref, '.json')
   end
@@ -102,51 +112,31 @@ class Codegen
   def capitalize(s)
     s.sub(/./,&:upcase)
   end
+
+  # Thank you very much rails!
+  # https://github.com/rails/rails/blob/v6.1.3.2/activesupport/lib/active_support/inflector/methods.rb#L92
+  def underscore(camel_cased_word)
+    return camel_cased_word unless /[A-Z-]/.match?(camel_cased_word)
+    word = camel_cased_word.gsub(/([A-Z\d]+)([A-Z][a-z])/, '\1_\2')
+    word.gsub!(/([a-z\d])([A-Z])/, '\1_\2')
+    word.tr!("-", "_")
+    word.downcase!
+    word
+  end
 end
 
 class TypeScript < Codegen
   def initialize(paths)
-    template = <<-EOF
-import { Type } from 'class-transformer'
-import 'reflect-metadata'
-
-<% @schemas.sort.each do |key, schema| -%>
-export class <%= class_name(key) %> {
-<% schema['properties'].each do |property_name, property| -%>
-<% ref = property['$ref'] || property['items'] && property['items']['$ref'] %>
-<% if ref -%>
-  @Type(() => <%= class_name(ref) %>)
-<% end -%>
-<% if (schema['required'] || []).index(property_name) -%>
-  <%= property_name %>: <%= type_for(class_name(key), property_name, property) %> = <%= default_value(class_name(key), property_name, property) %>
-<% else -%>
-  <%= property_name %>?: <%= type_for(class_name(key), property_name, property) %>
-<% end -%>
-<% end -%>
-}
-
-<% end -%>
-EOF
-
-    enum_template = <<-EOF
-export enum <%= enum[:name] %> {
-<% enum[:values].each do |value| -%>
-  <%= enum_constant(value) %> = '<%= value %>',
-<% end -%>
-}
-
-EOF
+    template = File.read("#{TEMPLATES_DIRECTORY}/typescript.ts.erb")
+    enum_template = File.read("#{TEMPLATES_DIRECTORY}/typescript.enum.ts.erb")
 
     language_type_by_schema_type = {
       'integer' => 'number',
       'string' => 'string',
       'boolean' => 'boolean',
     }
-    super(paths, template, enum_template, language_type_by_schema_type)
-  end
 
-  def property_type_from_ref(ref)
-    class_name(ref)
+    super(paths, template, enum_template, language_type_by_schema_type)
   end
 
   def array_type_for(type_name)
@@ -154,45 +144,60 @@ EOF
   end
 end
 
+class Ruby < Codegen
+  def initialize(paths, template_file_name: 'ruby.rb.erb')
+    template = File.read("#{TEMPLATES_DIRECTORY}/#{template_file_name}")
+    enum_template = File.read("#{TEMPLATES_DIRECTORY}/ruby.enum.rb.erb")
+
+    language_type_by_schema_type = {
+      'integer' => 'number',
+      'string' => 'string',
+      'boolean' => 'boolean',
+    }
+
+    super(paths, template, enum_template, language_type_by_schema_type)
+  end
+
+  def array_type_for(type_name)
+    "[]"
+  end
+
+  def default_value(parent_type_name, property_name, property)
+    if property['type'] == 'string'
+      if property['enum']
+        enum_type_name = type_for(parent_type_name, property_name, property)
+        "#{enum_type_name}::#{enum_constant(property['enum'][0])}"
+      else
+        "''"
+      end
+    elsif property['$ref']
+      type = type_for(parent_type_name, nil, property)
+      "#{type}.new"
+    else
+      super(parent_type_name, property_name, property)
+    end
+  end
+
+  def format_description(raw_description, indent_string: "    ")
+    return '' if raw_description.nil?
+
+    raw_description
+      .split("\n")
+      .map { |description_line| "# #{description_line}" }
+      .join("\n#{indent_string}")
+  end
+end
+
+class RubyDeserializers < Ruby
+  def initialize(paths)
+    super(paths, template_file_name: 'ruby_deserializers.rb.erb')
+  end
+end
+
 class Go < Codegen
   def initialize(paths)
-    template = <<-EOF
-package messages
-
-<% @schemas.sort.each do |key, schema| -%>
-type <%= class_name(key) %> struct {
-<% schema['properties'].each do |property_name, property| -%>
-<%
-type_name = type_for(class_name(key), property_name, property)
-required = (schema['required'] || []).index(property_name)
--%>
-  <%= capitalize(property_name) %> <%= type_name %> `json:"<%= property_name %><%= required ? '' : ',omitempty' %>"`
-<% end -%>
-}
-
-<% end -%>
-EOF
-
-    enum_template = <<-EOF
-type <%= enum[:name] %> string
-
-const(
-<% enum[:values].each do |value| -%>
-  <%= enum[:name] %>_<%= enum_constant(value) %> <%= enum[:name] %> = "<%= value %>"
-<% end -%>
-)
-
-func (e <%= enum[:name] %>) String() string {
-	switch e {
-<% enum[:values].each do |value| -%>
-  case <%= enum[:name] %>_<%= enum_constant(value) %>:
-    return "<%= value %>"
-<% end -%>
-	default:
-		panic("Bad enum value for <%= enum[:name] %>")
-	}
-}
-EOF
+    template = File.read("#{TEMPLATES_DIRECTORY}/go.go.erb")
+    enum_template = File.read("#{TEMPLATES_DIRECTORY}/go.enum.go.erb")
 
     language_type_by_schema_type = {
       'integer' => 'int64',
@@ -208,6 +213,32 @@ EOF
 
   def array_type_for(type_name)
     "[]#{type_name}"
+  end
+end
+
+class Markdown < Codegen
+  def initialize(paths)
+    template = File.read("#{TEMPLATES_DIRECTORY}/markdown.md.erb")
+    enum_template = File.read("#{TEMPLATES_DIRECTORY}/markdown.enum.md.erb")
+
+    language_type_by_schema_type = {
+      'integer' => 'integer',
+      'string' => 'string',
+      'boolean' => 'boolean',
+    }
+    super(paths, template, enum_template, language_type_by_schema_type)
+  end
+
+  def property_type_from_ref(ref)
+    "[#{class_name(ref)}](##{class_name(ref).downcase})"
+  end
+
+  def property_type_from_enum(enum)
+    "[#{enum}](##{enum.downcase})"
+  end
+
+  def array_type_for(type_name)
+    "#{type_name}[]"
   end
 end