+++ messages/ruby/lib/cucumber/messages.rb
@@ -1,6 +1,7 @@
 require 'cucumber/messages/ndjson_to_message_enumerator'
 require 'cucumber/messages/time_conversion'
 require 'cucumber/messages/id_generator'
+require 'cucumber/messages.deserializers'
 
 module Cucumber
   module Messages