+++ react/javascript/src/index.ts
@@ -5,20 +5,31 @@ import StatusIcon from './components/gherkin/StatusIcon'
 import filterByStatus from './filter/filterByStatus'
 import GherkinQueryContext from './GherkinQueryContext'
 import CucumberQueryContext from './CucumberQueryContext'
-import SearchQueryContext from './SearchQueryContext'
+import SearchQueryContext, {
+  WindowUrlApi,
+  SearchQueryUpdateFn,
+  SearchQueryProps,
+  searchFromURLParams,
+} from './SearchQueryContext'
 import EnvelopesQueryContext, { EnvelopesQuery } from './EnvelopesQueryContext'
 import MDG from './components/gherkin/MDG'
+import GherkinDocument from './components/gherkin/GherkinDocument'
 
 export {
   GherkinDocumentList,
   QueriesWrapper,
   GherkinQueryContext,
   CucumberQueryContext,
-  SearchQueryContext,
   StatusIcon,
   EnvelopesQueryContext,
   EnvelopesQuery,
   FilteredResults,
   filterByStatus,
+  SearchQueryContext,
+  WindowUrlApi,
+  SearchQueryUpdateFn,
+  SearchQueryProps,
+  searchFromURLParams,
   MDG,
+  GherkinDocument,
 }