+++ react/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+* Handle missing hook message when rendering hook failure
 * Stop the anchor link for the step from overlapping the status icon
 
 ## [16.2.0] - 2021-06-02